진행 전 리펙토링
