--- src/db/data.json
@@ -1,42 +1,0 @@
-[
-  {
-    "date": "2023-11-03T02:07:09.423Z",
-    "user": "Dr. Clint Christiansen",
-    "avatar": "https://cloudflare-ipfs.com/ipfs/Qmd3W5DuhgHirLHGVixi6V76LhCkZUz6pnFt5AJBiyvHye/avatar/36.jpg",
-    "comment": "카리나1 Vitae recusandae tenetur debitis impedit ut dolorem atque reprehenderit magnam. Cum dolor magnam commodi qui perferendis. Vel temporibus soluta. Eum delectus blanditiis. Neque dicta non quod ex. Maiores aspernatur fuga reprehenderit a magni eaque fuga voluptatum hic.",
-    "team": "두산 베어스",
-    "id": "1"
-  },
-  {
-    "date": "2023-11-02T23:13:18.491Z",
-    "user": "Chad Graham",
-    "avatar": "https://cloudflare-ipfs.com/ipfs/Qmd3W5DuhgHirLHGVixi6V76LhCkZUz6pnFt5AJBiyvHye/avatar/298.jpg",
-    "comment": "지젤1 Ipsam aspernatur nostrum eos unde velit molestiae dolorem. Tenetur ullam nostrum pariatur. Et in eos. Harum commodi ipsa quaerat aspernatur quod dignissimos quae quidem sapiente.",
-    "team": "삼성 라이온즈",
-    "id": "2"
-  },
-  {
-    "date": "2023-11-02T11:25:37.026Z",
-    "user": "Tommy Abshire",
-    "avatar": "https://cloudflare-ipfs.com/ipfs/Qmd3W5DuhgHirLHGVixi6V76LhCkZUz6pnFt5AJBiyvHye/avatar/646.jpg",
-    "comment": "윈터1 Itaque nihil quae neque itaque. Non a officiis ducimus nemo consectetur. Ducimus libero voluptatum consequuntur.",
-    "team": "한화 이글즈",
-    "id": "3"
-  },
-  {
-    "date": "2023-11-02T16:06:34.150Z",
-    "user": "Max Mayert",
-    "avatar": "https://cloudflare-ipfs.com/ipfs/Qmd3W5DuhgHirLHGVixi6V76LhCkZUz6pnFt5AJBiyvHye/avatar/37.jpg",
-    "comment": "닝닝1 Sint qui eligendi repudiandae placeat maiores repudiandae assumenda repudiandae. Distinctio commodi iste. Qui architecto iusto.",
-    "team": "두산 베어스",
-    "id": "4"
-  },
-  {
-    "date": "2023-11-03T05:40:17.575Z",
-    "user": "Olga Christiansen",
-    "avatar": "https://cloudflare-ipfs.com/ipfs/Qmd3W5DuhgHirLHGVixi6V76LhCkZUz6pnFt5AJBiyvHye/avatar/720.jpg",
-    "comment": "카리나2 Molestiae saepe reiciendis saepe natus quo occaecati. Vel vero illum quo. Ducimus maiores porro optio illum officia nam. Cum possimus aut consequatur eaque libero ad nihil pariatur officiis.",
-    "team": "LG 트윈스",
-    "id": "5"
-  }
-]
